+++ src/main/java/com/github/hcsp/inheritance/Person.java
@@ -1,0 +1,14 @@
+package com.github.hcsp.inheritance;
+
+public class Person extends Animal {
+    private final int age;
+
+    public Person(String name,int age) {
+        super(name);
+        this.age = age;
+    }
+
+    public void sayMyAge() {
+        System.out.println("我的年龄是" + age);
+    }
+}
